add ocean fixes for cnrm models
