+++ esmvalcore/cmor/_fixes/cmip6/cnrm_esm2_1.py
@@ -1,12 +1,9 @@
 """Fixes for CNRM-ESM2-1 model."""
-from ..fix import Fix
-from ..shared import (fix_ocean_depth_coord)
-
-
 from .cnrm_cm6_1 import Cl as BaseCl
 from .cnrm_cm6_1 import Clcalipso as BaseClcalipso
 from .cnrm_cm6_1 import Cli as BaseCli
 from .cnrm_cm6_1 import Clw as BaseClw
+from .cnrm_cm6_1 import Omon as BaseOmon
 
 
 Cl = BaseCl
@@ -21,25 +18,4 @@ Cli = BaseCli
 Clw = BaseClw
 
 
-class Omon(Fix):
-    """Fixes for ocean variables."""
-
-    def fix_metadata(self, cubes):
-        """Fix ocean depth coordinate.
-
-        Parameters
-        ----------
-        cubes: iris CubeList
-            List of cubes to fix
-
-        Returns
-        -------
-        iris.cube.CubeList
-
-        """
-        for cube in cubes:
-            if cube.coords(axis='Z'):
-                z_coord = cube.coord(axis='Z')
-                if z_coord.standard_name is None:
-                    fix_ocean_depth_coord(cube)
-        return cubes
+Omon = BaseOmon
